Found a better version of the scratch-paper/doodle code. It's a Doodle.java file that uses a DoodleView.java (I'll have to check if the view one uploaded - they should both have the same date).
I didn't plug it into the main program - I'm concerned I might break
something. James, can you take a look at that? I know it needs a method
to create the Doodle as a new intent and start the activity, but adding
the xml for the action bar button is where I'm stumped.